refactor: move GameController Prompt tests to test class with Status tests and leave CheckCurrentLocation in its own test class
